pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django-ai-lens"
version = "0.1.0"
description = "Natural language queries for Django models, powered by AI"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "Django AI Lens" }]
keywords = ["django", "ai", "llm", "gemini", "natural-language", "orm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "Django>=4.0",
    "google-genai>=1.0",
    "pydantic>=2.0",
    "langchain-core>=0.2",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-django",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["django_ai_lens*"]
